+++ pageserver/src/branches.rs
@@ -7,7 +7,7 @@
 use anyhow::{anyhow, bail, Context, Result};
 use bytes::Bytes;
 use fs::File;
-use postgres_ffi::{pg_constants, xlog_utils};
+use postgres_ffi::{controlfile_utils, pg_constants, xlog_utils};
 use rand::Rng;
 use serde::{Deserialize, Serialize};
 use std::env;
@@ -80,7 +80,7 @@ pub fn init_repo(conf: &'static PageServerConf, repo_dir: &Path) -> Result<()> {
 
     // Read control file to extract the LSN and system id
     let controlfile_path = tmppath.join("global").join("pg_control");
-    let controlfile = postgres_ffi::decode_pg_control(Bytes::from(fs::read(controlfile_path)?))?;
+    let controlfile = controlfile_utils::decode_pg_control(Bytes::from(fs::read(controlfile_path)?))?;
     // let systemid = controlfile.system_identifier;
     let lsn = controlfile.checkPoint;
     let lsnstr = format!("{:016X}", lsn);
@@ -212,7 +212,7 @@ pub(crate) fn get_system_id(conf: &PageServerConf) -> Result<u64> {
 
     let (_, main_snap_dir) = find_latest_snapshot(conf, *main_tli)?;
     let controlfile_path = main_snap_dir.join("global").join("pg_control");
-    let controlfile = postgres_ffi::decode_pg_control(Bytes::from(fs::read(controlfile_path)?))?;
+    let controlfile = controlfile_utils::decode_pg_control(Bytes::from(fs::read(controlfile_path)?))?;
     Ok(controlfile.system_identifier)
 }
 
@@ -356,13 +356,13 @@ fn force_crash_recovery(datadir: &Path) -> Result<()> {
     // Read in the control file
     let controlfilepath = datadir.to_path_buf().join("global").join("pg_control");
     let mut controlfile =
-        postgres_ffi::decode_pg_control(Bytes::from(fs::read(controlfilepath.as_path())?))?;
+        controlfile_utils::decode_pg_control(Bytes::from(fs::read(controlfilepath.as_path())?))?;
 
     controlfile.state = postgres_ffi::DBState_DB_IN_PRODUCTION;
 
     fs::write(
         controlfilepath.as_path(),
-        postgres_ffi::encode_pg_control(controlfile),
+        controlfile_utils::encode_pg_control(controlfile),
     )?;
 
     Ok(())
